requirements.txt
numpy
pandas
scikit-learn
bokeh
ipywidgets
streamlit
pillow
pytesseract

